Done adding support for NMA NDD

+++ include/ANA/TextUtils.hpp
@@ -13,6 +13,6 @@ void write_vector(std::vector<double> vec, std::string const &filename);
 auto slurp(std::string const &filename) -> std::unique_ptr<char[]>;
 
 //
-auto guess_format(std::string_view texto) -> std::tuple<int, int, int>;
+auto guess_format(std::string_view texto) -> std::tuple<size_t, size_t, size_t>;
 }
 #endif // _H
